chore: remove pnpm lock from gitignore

docs: improve docs + add installation steps

batch features: improve styling, ux control flow, and component impl

- refactor styles
- update bacgkround padding
- use monospace font
- add standard message compoennts
- reduce inline styling

feat: add error page

feat: break fork syncing and account impersonation into a separate
method

feat: add method for 'syncing' controller with server

the server and controller often times run out of sync (server is
restarted, accounts aren't impersonated, etc). it's useful to have the
inpage scripts and / or ui scripts be able to prompt the controller to
resync itself

fix: make sure neptune provider "beats" metamask provider. improve docs

docs: improve inline comments

feat: improve ux flow, allow fork deletion

target ux flow:

- when server is off, report it in ui
- when server is on, make sure there's a fork active, and make sure the
  ui knows about it

fix: remove unneeded key on FormControl

feat: point README install section to extension installation instructions

docs: improve build / installation docs

feat: add and use delayed spinner to avoid spinner flashing

fix: move controller side effects into setters, not listeners

the background script does not receive change messages that it sends
itself. we therefore cannot rely on onChange listeners registered in the
constructor

fix: remove "Fork Mainnet" confusion on forks page

- the ForkMainnet components forks mainnet with fork name derived from
  mainnet block
- the previously "Blank Fork" button would also fork mainnent, but with
  the given name

this commimt remove the "ForkMainnet" component from the forks page, and
changes the langauge of the named fork creation from "Blank Fork" to
"Fork Mainnet"

fix: add padding to Centered background component

fix: correct spelling on extension/docs/install-on-chrome.md

+++ extension/src/shared/mutations.ts
@@ -7,15 +7,17 @@ type CreateForkParams = {
   name: string
   from?: Fork | null
   baseUrl?: URL | string | null | void
+  providerRpcUrl?: URL | string | void
 }
 
 export const createFork = async ({
   from,
   name,
   baseUrl: _baseUrl,
+  providerRpcUrl: _providerRpcUrl,
 }: CreateForkParams): Promise<Fork> => {
   const baseUrl = _baseUrl ?? (await msg.baseUrl.get())
-  const providerRpcUrl = await msg.providerRpcUrl.get()
+  const providerRpcUrl = _providerRpcUrl ?? (await msg.providerRpcUrl.get())
 
   const config = from
     ? {
@@ -50,25 +52,42 @@ export const createFork = async ({
   }
 }
 
-export const deleteFork = async (fork: Fork) => {
+export const deleteActiveFork = async () => {
+  const fork = await msg.fork.get()
   const baseUrl = await msg.baseUrl.get()
-  await fetch(url.fork(baseUrl, fork), {
+
+  if (!fork) throw new Error('No active fork')
+
+  const json = await fetch(url.fork(baseUrl, fork), {
     method: 'DELETE',
-  })
+  }).then(res => res.json())
+
+  if (!json.ok) {
+    if (json.error.Readonly) {
+      throw new Error(
+        `Cannot delete fork ${fork.name} because it is readonly. ` +
+          `It likely has child forks depending on it.`,
+      )
+    }
+
+    throw new Error('Failed to delete fork')
+  }
+
+  await msg.sync.sync()
 }
 
 export const forkMainnetLatest = async (
   baseUrl?: URL | string | void,
+  providerRpcUrl?: URL | string | void,
 ): Promise<Fork> => {
-  const providerRpcUrl = await msg.providerRpcUrl.get()
+  const _providerRpcUrl = providerRpcUrl ?? (await msg.providerRpcUrl.get())
 
-  if (!providerRpcUrl) {
-    throw new Error('No provider rpc url set')
-  }
+  if (!_providerRpcUrl) throw new Error('No provider rpc url set')
 
-  const block = await latestBlock(providerRpcUrl)
+  const block = await latestBlock(_providerRpcUrl)
   const name = `Mainnet @ block ${block}`
-  return createFork({ name, baseUrl })
+
+  return createFork({ name, baseUrl, providerRpcUrl })
 }
 
 type BacktrackOptions =
